+++ src/_worker.ts
@@ -26,9 +26,9 @@ router
   .post(`${API.Consts.API_LATEST}/customer/discount`, Customer.discount)
   .post(`${API.Consts.API_LATEST}/shop/redact`, Shop.redact)
   .get(`${API.Consts.API_LATEST}/shop/uninstall`, Shop.uninstall)
+  .post(`${API.Consts.API_LATEST}/shop/image`, Shop.image)
   .post(`${API.Consts.API_LATEST}/discount`, Discount.create)
   .get(`${API.Consts.API_LATEST}/discount/:id`, Discount.get)
-  .post(`${API.Consts.API_LATEST}/upload/stage`, Discount.stagedUpload)
   .all('*', Shop.ui);
 
 export default {
